Rename ViewController classes and move logic accordingly.
Task #19

TODO: Check instance holding of new ViewController
