+++ de.tu_bs.cs.isf.e4cf.parts.project_explorer/project_explorer.e4xmi
@@ -2,7 +2,7 @@
 <fragment:ModelFragments xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:basic="http://www.eclipse.org/ui/2010/UIModel/application/ui/basic" xmlns:commands="http://www.eclipse.org/ui/2010/UIModel/application/commands" xmlns:fragment="http://www.eclipse.org/ui/2010/UIModel/fragment" xmlns:menu="http://www.eclipse.org/ui/2010/UIModel/application/ui/menu" xmi:id="_RuDIkPfoEeiagocgPMLR0w">
   <imports xsi:type="commands:Command" xmi:id="_RuDIkffoEeiagocgPMLR0w" elementId="de.tu_bs.cs.isf.familymining.ppu_iec.rcp_e4.command.open_file_command" commandName="open_file_command" description="Open a file in the editor"/>
   <fragments xsi:type="fragment:StringModelFragment" xmi:id="_RuDIkvfoEeiagocgPMLR0w" featurename="children" parentElementId="de.tu_bs.cs.isf.familymining.ppu_iec.rcp_e4.left_window" positionInList="index:0">
-    <elements xsi:type="basic:Part" xmi:id="_RuDIk_foEeiagocgPMLR0w" elementId="de.tu_bs.cs.isf.e4cf.parts.project_explorer" contributionURI="bundleclass://de.tu_bs.cs.isf.e4cf.parts.project_explorer/de.tu_bs.cs.isf.e4cf.parts.project_explorer.ProjectExplorerViewController" label="Project Explorer" iconURI="platform:/plugin/de.tu_bs.cs.isf.e4cf.core/icons/Explorer_View/items/project_explore_24.png">
+    <elements xsi:type="basic:Part" xmi:id="_RuDIk_foEeiagocgPMLR0w" elementId="de.tu_bs.cs.isf.e4cf.parts.project_explorer" contributionURI="bundleclass://de.tu_bs.cs.isf.e4cf.parts.project_explorer/de.tu_bs.cs.isf.e4cf.parts.project_explorer.ProjectExplorerViewAdapter" label="Project Explorer" iconURI="platform:/plugin/de.tu_bs.cs.isf.e4cf.core/icons/Explorer_View/items/project_explore_24.png">
       <menus xsi:type="menu:PopupMenu" xmi:id="_RuDIlffoEeiagocgPMLR0w" elementId="de.tu_bs.cs.isf.e4cf.parts.project_explorer.context_menu" label="project_explorer_context_menu">
         <children xsi:type="menu:DirectMenuItem" xmi:id="_RuDIlvfoEeiagocgPMLR0w" elementId="de.tu_bs.cs.isf.familymining.ppu_iec.rcp_e4.directmenuitem.new_folder" label="New Folder" iconURI="platform:/plugin/de.tu_bs.cs.isf.e4cf.core/icons/Explorer_View/items/project16.png" contributionURI="bundleclass://de.tu_bs.cs.isf.e4cf.parts.project_explorer/de.tu_bs.cs.isf.e4cf.parts.project_explorer.handlers.NewFolderHandler"/>
         <children xsi:type="menu:DirectMenuItem" xmi:id="_RuDIl_foEeiagocgPMLR0w" elementId="de.tu_bs.cs.isf.familymining.ppu_iec.rcp_e4.directmenuitem.import_files" label="Import Files" iconURI="platform:/plugin/de.tu_bs.cs.isf.e4cf.core/icons/Explorer_View/items/file16.png" tooltip="Import one or more files into a selected folder" contributionURI="bundleclass://de.tu_bs.cs.isf.e4cf.parts.project_explorer/de.tu_bs.cs.isf.e4cf.parts.project_explorer.handlers.FileImportHandler"/>
